Extract timestamp synchronization logic into TimestampSynchronizer
